cpu-minor: Add control instruction stats to Minor CPU
Add control/branch instructions committed stat to Minor CPU. The stats
can be found in board.processor.cores.core.commitedControl_0 in
stats.txt. The stats counted are Call, Return, DirectCtrl,
IndirectCtrl, CondCtrl, UncondCtrl, and Control. The last one tracks
the total control/branch instructions committed.
